Fix models.
